fixed import cycle not allowed in test
