+++ middlewares/config_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/gorilla/mux"
 	"github.com/prest/prest/adapters/postgres"
 	"github.com/prest/prest/config"
-	"github.com/prest/prest/config/router"
 	"github.com/prest/prest/controllers"
+	"github.com/prest/prest/router"
 	"github.com/urfave/negroni"
 )
